+++ packages/dataviews/src/components/dataform-controls/index.tsx
@@ -22,7 +22,6 @@ import select from './select';
 import text from './text';
 import toggle from './toggle';
 import textarea from './textarea';
-import richtext from './richtext';
 import toggleGroup from './toggle-group';
 import array from './array';
 import color from './color';
@@ -52,7 +51,6 @@ const FORM_CONTROLS: FormControls = {
 	text,
 	toggle,
 	textarea,
-	richtext,
 	toggleGroup,
 };